+++ e2e/src/app.po.ts
@@ -6,7 +6,7 @@ export class AppPage {
   }
 
   getTitleText() {
-    return element(by.css('app-root h1')).getText() as Promise<string>;
+    return element(by.css('body h1')).getText() as Promise<string>;
   }
 
   getIntegerInput() {